refactor and rename ContractOptions to ContractTabs
